Factor out RunnerCore
